requirements.txt
streamlit
streamlit-autorefresh
mysql-connector-python
pandas
requests